+++ KittsMenuSystem/Examples/AccessExmaple.cs
@@ -1,5 +1,5 @@
 using KittsMenuSystem.Features.Menus;
-using KittsMenuSystem.Features.Wrappers;
+using KittsMenuSystem.Features.Settings;
 using System.Collections.Generic;
 
 namespace KittsMenuSystem.Examples;