+++ KittsMenuSystem/Examples/DemoExample.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using KittsMenuSystem.Features.Menus;
-using KittsMenuSystem.Features.Wrappers;
+using KittsMenuSystem.Features.Settings;
 using UserSettings.ServerSpecific;
 
 namespace KittsMenuSystem.Examples;